add mechanism for preserving config files
Add a simple way of preserving files in /etc and /etc/conf.d.

In order to preserve these files, the user only needs to copy them to
/mnt/nv/tiny/conf directory.  After reboot, rc will symlink from /etc or
/etc/conf.d to the new copies in /mnt/nv/tiny/conf.

Editing the file in /etc or in /mnt/nv/tiny/conf after the next reboot
will have the same effect.

To restore originals from SquashFS it is sufficient to delete the
copies.

Change-Id: Ie0df2c169bdc37e00688e0e28e7800759990b6cd
Reviewed-on: http://git-master/r/1133613
Tested-by: Chris Dragan <kdragan@nvidia.com>
Reviewed-by: Lael Jones <lajones@nvidia.com>
Reviewed-by: Chris Dragan <kdragan@nvidia.com>